+++ EXTI/Core/Inc/functions.h
@@ -1,0 +1,14 @@
+/*
+ * functions.h
+ *
+ *  Created on: Mar 5, 2020
+ *      Author: rodrigocelmer
+ */
+
+#ifndef INC_FUNCTIONS_H_
+#define INC_FUNCTIONS_H_
+
+void Blink_Red_Led (void);		//blinks the red led with the frequency set using the user button
+void Change_Frequency (void);	//changes the led frequency every user button click
+
+#endif /* INC_FUNCTIONS_H_ */
