+++ EXTI/Core/Inc/init.h
@@ -1,0 +1,14 @@
+/*
+ * init.h
+ *
+ *  Created on: Mar 5, 2020
+ *      Author: rodrigocelmer
+ */
+
+#ifndef INC_INIT_H_
+#define INC_INIT_H_
+
+void GPIO_Init (void);
+void TIM6_Init (void);	//Prescaler output: 1MHz/1 us
+
+#endif /* INC_INIT_H_ */
